Read input into EventChannel
Since NetConnection does not in fact duplicate events for the client
side, we read our input, tagged with who's it is, into a shared channel.

This might be a serious problem which is that EventChannel is supposedly
"single-producer", but we're writing to it from net_recieve AND input.
More research to be done